wip formulaire ajout film avec genres

+++ src/main/java/fr/eni/filmotheque/bll/FilmsService.java
@@ -7,6 +7,6 @@ import fr.eni.filmotheque.bo.Film;
 public interface FilmsService 
 {
 	public List<Film> getFilms();
-	public Film getFilmById(long id);
+	public Film getFilmById(Integer id);
 	public void insertFilm(Film film);
 }
