+++ src/main/java/fr/eni/filmotheque/bo/Genre.java
@@ -2,22 +2,22 @@ package fr.eni.filmotheque.bo;
 
 public class Genre {
 
-	private long id;
+	private Integer id;
 	private String name;
 	
 	public Genre() {
 	}
 
-	public Genre(long id, String name) {
+	public Genre(Integer id, String name) {
 		this.id = id;
 		this.name = name;
 	}
 	
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
